backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
websockets==12.0
pydantic==2.5.0
torch>=2.0.0
lerobot
huggingface_hub
python-dotenv==1.0.0
asyncio
aiofiles==23.2.1
redis==5.0.1